+++ functional_data/lib/src/annotations.dart
@@ -14,11 +14,11 @@ class Ignore implements Equality {
   const Ignore();
 
   @override
-  bool equals(_, __) => true;
+  bool equals(dynamic _, dynamic __) => true;
 
   @override
   int hash(_) => 0;
 
   @override
-  bool isValidKey(Object o) => true;
+  bool isValidKey(Object? o) => true;
 }